finishing touches and update landing page

+++ routes/todos.js
@@ -1,5 +1,5 @@
 const {Router} = require('express')
-const { getTodosByUserId, createTodo, updateTodo, deleteTodo } = require('../db')
+const { getTodosByUserId, createTodo, updateTodo, deleteTodo, getTodoById } = require('../db')
 const verifiers = require('../verifiers')
 const {asyncWrapper} = require("../util")
 
@@ -25,6 +25,12 @@ todosRouter.patch("/:id",
         verifiers.updateTodo,
         verifiers.validCurrentUser,
         asyncWrapper(async (req, res) => {
+    const todoCheck = await getTodoById(req.body.id);
+
+    if (todoCheck.userId !== req.session.user.id) {
+        return res.status(401).json({message: "unauthorized"})
+    }
+
     const todo = await updateTodo(req.params.id, req.body)
     res.json(todo)
 }))
@@ -33,6 +39,12 @@ todosRouter.delete("/:id",
         verifiers.idParam,
         verifiers.validCurrentUser,
         asyncWrapper(async (req, res) => {
+    const todoCheck = await getTodoById(req.params.id);
+
+    if (todoCheck.userId !== req.session.user.id) {
+        return res.status(401).json({message: "unauthorized"})
+    }
+
     await deleteTodo(req.params.id)
     res.json({message: "successfully deleted todo"})
 }))
